requirements.txt
torch
torchvision
numpy
opencv-python
matplotlib
scikit-learn
tqdm
pandas
pillow
facenet-pytorch

